+++ src/main/java/cz/uhk/pro2/tulipani/domain/repository/AppUserRepository.java
@@ -4,7 +4,7 @@ import cz.uhk.pro2.tulipani.domain.entity.AppUser;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppUserRepository extends JpaRepository<AppUser, Long> {
+public interface AppUserRepository extends JpaRepository<AppUser, Integer> {
 
     Optional<AppUser> findByAuthId(String authId);
 }
